NavigationFilters use generic Selenium objects; added abstract SeleniumProtocol class and JBrowser extension. Moved main test method to top AbstractHttpProtocol and use it in httpclient and jbrowser classes
